better HTTP verb deducation
